Big refacgtoring: Tools are now classes, each living in their own module
